refactor: isolate LSP options from lsp transport; create abstract transport
